front

+++ services/Auth/auth_logic.py
@@ -96,10 +96,7 @@ def save_refresh_token(
 ):
     token_hash = get_password_hash(refresh_token)
 
-    db.query(RefreshToken).filter(
-        RefreshToken.user_id == user_id,
-        RefreshToken.device_info == device_info
-    ).delete()
+    db.query(RefreshToken).filter(RefreshToken.user_id == user_id).delete()
 
     db.add(
         RefreshToken(
